Fix compute bugs causing NaN timestamps and PostgreSQL errors
Fixes 4 critical compute bugs identified in skill validation:

1. NaN timestamp bug (4 locations):
   - waterfallAnalysisTool: Fix timeWindows property access
   - topDealsInMotionTool: Fix timeWindows property access
   - velocityBenchmarksTool: Fix timeWindows property access
   - repScorecardComputeTool: Fix timeWindows property access
   Root cause: Reading from non-existent properties like timeWindows.analysisStart
   instead of timeWindows.analysisRange.start, causing new Date(undefined) which
   produces Invalid Date and renders as "0NaN-NaN-NaN" in output.

2. PostgreSQL ROUND() type mismatch (2 locations):
   - getStageConversionRates: Add ::NUMERIC cast before ROUND()
   - getAverageTimeInStage: Add ::NUMERIC cast before ROUND()
   Root cause: PostgreSQL doesn't support ROUND(double precision, 2), requires
   ROUND(value::NUMERIC, 2) instead.

3. waterfallDeltas crash:
   - Add null check for current.stages and previous.stages before .map()
   Root cause: Attempting .map() on undefined when upstream returns no data.

These fixes resolve PARTIAL status for pipeline-waterfall and rep-scorecard skills.

+++ server/analysis/stage-history-queries.ts
@@ -193,7 +193,7 @@ export async function getStageConversionRates(
       COUNT(*) AS transition_count,
       CASE
         WHEN AVG(dsh.duration_in_previous_stage_ms) IS NOT NULL
-        THEN ROUND(AVG(dsh.duration_in_previous_stage_ms) / (1000.0 * 60 * 60 * 24), 2)
+        THEN ROUND((AVG(dsh.duration_in_previous_stage_ms) / (1000.0 * 60 * 60 * 24))::NUMERIC, 2)
         ELSE NULL
       END AS avg_duration_days
     FROM deal_stage_history dsh
@@ -410,10 +410,10 @@ export async function getAverageTimeInStage(
   }>(
     `SELECT
       dsh.to_stage_normalized AS stage,
-      ROUND(AVG(dsh.duration_in_previous_stage_ms) / (1000.0 * 60 * 60 * 24), 2) AS avg_days,
-      ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (
+      ROUND((AVG(dsh.duration_in_previous_stage_ms) / (1000.0 * 60 * 60 * 24))::NUMERIC, 2) AS avg_days,
+      ROUND((PERCENTILE_CONT(0.5) WITHIN GROUP (
         ORDER BY dsh.duration_in_previous_stage_ms
-      ) / (1000.0 * 60 * 60 * 24), 2) AS median_days,
+      ) / (1000.0 * 60 * 60 * 24))::NUMERIC, 2) AS median_days,
       COUNT(*) AS deal_count
     FROM deal_stage_history dsh
     WHERE dsh.workspace_id = $1
